调低修炼速度，调高恢复速度，调高恢复上限，将原版恢复满血调整为恢复满护盾

+++ Code/CW_ActorStatus.cs
@@ -8,8 +8,8 @@ namespace Cultivation_Way
 {
 	public class CW_ActorStatus
 	{
-		public int wakan;
-		public int shield;
+		public float wakan;
+		public float shield;
 		public float culti_velo;
 		public bool can_culti;
 		public float health_level;
